Fix API Key flyout double submit (#167468)
## Summary

Closes https://github.com/elastic/kibana/issues/163314

## Fixes

- Removed an extra EuiPortal that was not needed as EuiFlyout adds a
Portal
- Inverted control of the form by wrapping the flyout content in the
form. Prevents multiple submits by using traditional form controls and
button type as submit.

## Release Notes:

- Fixes issue with multiple API keys being created if the form is
submitted using the enter key fired multiple times in quick succession.
https://github.com/elastic/kibana/issues/163314

---------

(cherry picked from commit e3d9f3d62e403f56b9d3a553338a3c5201edc021)
